requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.3.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
python-decouple==3.8
Pillow==11.0.0
